+++ packages/gatsby-acme-admin-theme/theme/index.js
@@ -6,6 +6,7 @@ const MuiAppBar = require('./mui/muiComponentsOverrides/mui-app-bar')
 const MuiPopover = require('./mui/muiComponentsOverrides/mui-popover')
 const MuiButton = require('./mui/muiComponentsOverrides/mui-button')
 const MuiInput = require('./mui/muiComponentsOverrides/mui-input')
+const MuiGrid = require('./mui/muiComponentsOverrides/mui-grid')
 
 // Section below are theme key word specific overrides https://material-ui.com/customization/default-theme/
 const palette = require('./mui/palette')
@@ -18,6 +19,7 @@ const footer = require('./custom/footer')
 const layout = require('./custom/layout')
 const searchInput = require('./custom/search-input')
 const secondaryAppBar = require('./custom/secondary-app-bar')
+const form = require('./custom/form')
 
 exports.muiTheme = {
   overrides: {
@@ -29,6 +31,7 @@ exports.muiTheme = {
     MuiPopover,
     MuiButton,
     MuiInput,
+    MuiGrid,
   },
   palette,
   typography,
@@ -38,6 +41,7 @@ exports.muiTheme = {
   layout,
   searchInput,
   secondaryAppBar,
+  form,
 }
 
 exports.googleFonts = ['Titillium Web:400,700,900']