+++ packages/gatsby-plugin-orion-view/src/components/UserLogin/index.stories.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { jsxDecorator } from 'storybook-addon-jsx'
+
+import UserLogin from '.'
+
+storiesOf('User Login', module)
+  .addDecorator(jsxDecorator)
+  .add('default', () => <UserLogin />)