+++ packages/gatsby-plugin-orion-view/src/components/UserRegistration/index.stories.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { jsxDecorator } from 'storybook-addon-jsx'
+
+import UserRegistration from '.'
+
+storiesOf('User Registration', module)
+  .addDecorator(jsxDecorator)
+  .add('default', () => <UserRegistration />)