+++ packages/gatsby-acme-admin-theme/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/nearform/orion/compare/v2.13.1...v2.14.0) (2020-02-24)
+
+
+### Features
+
+* 🎸 Basic UploadButton added ([e2029a8](https://github.com/nearform/orion/commit/e2029a8741cf7f26ee52a1ea9a47395d6a3289fd))
+* 🎸 Created Sign in component for orion-view ([f0e0b32](https://github.com/nearform/orion/commit/f0e0b32d840b0aed4c3e8bbdcf9412153bb09b86))
+* 🎸 Finished form for artifle creation ([b9b9b63](https://github.com/nearform/orion/commit/b9b9b63b2bc49cfca6374d79444c16c014ea4d51))
+
+
+
+
+
 ## [2.13.1](https://github.com/nearform/orion/compare/v2.13.0...v2.13.1) (2020-02-21)